+++ src/BlazorWebFormsComponents/Enums/ListViewCancelMode.cs
@@ -1,0 +1,8 @@
+namespace BlazorWebFormsComponents.Enums
+{
+	public enum ListViewCancelMode
+	{
+		CancelingEdit = 0,
+		CancelingInsert = 1
+	}
+}