+++ src/BlazorWebFormsComponents/Enums/LoginTextLayout.cs
@@ -1,0 +1,7 @@
+namespace BlazorWebFormsComponents.Enums;
+
+public enum LoginTextLayout
+{
+	TextOnLeft = 0,
+	TextOnTop = 1
+}