+++ src/BlazorWebFormsComponents/GridViewPagerStyle.razor.cs
@@ -1,0 +1,20 @@
+using BlazorWebFormsComponents.Interfaces;
+using Microsoft.AspNetCore.Components;
+
+namespace BlazorWebFormsComponents
+{
+	public partial class GridViewPagerStyle : UiTableItemStyle
+	{
+		[CascadingParameter(Name = "ParentGridView")]
+		protected IGridViewStyleContainer ParentGridView { get; set; }
+
+		protected override void OnInitialized()
+		{
+			if (ParentGridView != null)
+			{
+				theStyle = ParentGridView.PagerStyle;
+			}
+			base.OnInitialized();
+		}
+	}
+}